+++ src/index.ts
@@ -1,15 +1,6 @@
-export class TSMap<K,V> {
+export class TSMap<K, V> {
 
-    public length:number;
-
-    /**
-     * Used to hold any array of the map items.
-     * 
-     * @internal
-     * @type {(Array<Array<K|V>>)}
-     * @memberOf TSMap
-     */
-    private _items:Array<Array<K|V>>;
+    public length: number;
 
     /**
      * Used to hold an array of keys in the map
@@ -18,7 +9,7 @@ export class TSMap<K,V> {
      * @type {Array<K>}
      * @memberOf TSMap
      */
-    private _keys:Array<K>;
+    private _keys: K[];
 
     /**
      * Used to hold an array of values in the map
@@ -27,52 +18,69 @@ export class TSMap<K,V> {
      * @type {Array<V>}
      * @memberOf TSMap
      */
-    private _values:Array<V>;
+    private _values: V[];
 
-    constructor(inputMap?:Array<Array<K|V>>) {
+    constructor(inputMap?: (K | V)[][]) {
         let t = this;
 
-        t._items = [];
         t._keys = [];
         t._values = [];
         t.length = 0;
 
-        if(inputMap) {
-            inputMap.forEach((v,k) => {
-                t.set(<K> v[0],<V> v[1]);
+        if (inputMap) {
+            inputMap.forEach((v, k) => {
+                t.set(v[0] as K, v[1] as V);
             });
         }
     }
 
     /**
-     * Converts a JSON object to an map.
+     * Convert a JSON object to a map.
      * 
-     * @param {*} jsonObject
-     * 
-     * @memberOf TSMap
+     * @param {*} jsonObject JSON object to convert
+     * @param {boolean} [convertObjs] convert nested objects to maps
+     * @returns {TSMap<K, V>} 
+     * @memberof TSMap
      */
-    public fromJSON(jsonObject:any): TSMap<K,V> {
-        for (let property in jsonObject) {  
-            if (jsonObject.hasOwnProperty(property)) {
-                this.set(<any> property, jsonObject[property]);
-            }
-        }        
-        return this;
-    }
+    public fromJSON(jsonObject: any, convertObjs?: boolean): this {
+        let t = this;
 
+        const setProperty = (value: any): any => {
+            if (value !== null && typeof value === 'object' && convertObjs) return new TSMap<any,any>().fromJSON(value, true);
+            if (Array.isArray(value) && convertObjs) return value.map(v => setProperty(v));
+            return value;
+        }
+
+        Object.keys(jsonObject).forEach((property) => {
+            if (jsonObject.hasOwnProperty(property)) {
+                t.set(property as any, setProperty(jsonObject[property]));
+            }
+        });
+        return t;
+    }
 
     /**
      * Outputs the contents of the map to a JSON object
      * 
-     * @returns {Object}
-     * 
-     * @memberOf TSMap
+     * @returns {{[key: string]: V}} 
+     * @memberof TSMap
      */
-    public toJSON():Object {
+    public toJSON(): {[key: string]: V} {
         let obj = {};
         let t = this;
+
+        const getValue = (value: any): any => {
+            if (value instanceof TSMap) {
+                return value.toJSON();
+            } else if (Array.isArray(value)) {
+                return value.map(v => getValue(v));
+            } else {
+                return value;
+            }
+        }
+
         t.keys().forEach((k) => {
-            obj[String(k)] = t.get(k);
+            obj[String(k)] = getValue(t.get(k));
         });
         return obj;
     }
@@ -84,8 +92,8 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public entries():Array<Array<K|V>> {
-        return [].slice.call(this._items);
+    public entries(): (K|V)[][] {
+        return [].slice.call(this.keys().map(k => [k, this.get(k)]));
     }
 
     /**
@@ -95,7 +103,7 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public keys():Array<K> {
+    public keys(): K[] {
         return [].slice.call(this._keys);
     }
 
@@ -106,7 +114,7 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public values():Array<V> {
+    public values(): V[] {
         return [].slice.call(this._values);
     }
 
@@ -118,7 +126,7 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public has(key:K):Boolean {
+    public has(key: K): boolean {
         return this._keys.indexOf(key) > -1;
     }
 
@@ -130,10 +138,31 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public get(key:K):V {
+    public get(key: K): V {
         let i = this._keys.indexOf(key);
-        return i > -1 ? this._values[i] : undefined;        
+        return i > -1 ? this._values[i] : undefined;
     }
+
+    /**
+     * Safely retrieve a deeply nested property.
+     * 
+     * @param {K[]} path 
+     * @returns {V} 
+     * 
+     * @memberOf TSMap
+     */
+    public deepGet(path: K[]): V {
+        if (!path || !path.length) return null;
+
+        const recursiveGet = (obj: any, path: K[]) => {
+            if (obj === undefined || obj === null) return null;
+            if (!path.length) return obj;
+            return recursiveGet(obj instanceof TSMap ? obj.get(path[0]) : obj[path[0]], path.slice(1));
+        }
+
+        return recursiveGet(this.get(path[0]), path.slice(1));
+    }
+
 
     /**
      * Set a specific item in the map given it's key, automatically adds new items as needed. 
@@ -144,19 +173,17 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public set(key:K, value:V): TSMap<K,V> {
+    public set(key: K, value: V): this {
         let t = this;
         // check if key exists and overwrite
         let i = this._keys.indexOf(key);
         if (i > -1) {
-            t._items[i][1] = value;
             t._values[i] = value;
         } else {
-            t._items.push([key, value]);
             t._keys.push(key);
             t._values.push(value);
+            t.length = t._values.length;
         }
-        t.length = t.size();
         return this;
     }
 
@@ -167,8 +194,8 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public size():number {
-        return this._items.length;
+    public size(): number {
+        return this.length;
     }
 
 
@@ -179,9 +206,9 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public clear(): TSMap<K,V> {
+    public clear(): this {
         let t = this;
-        t._keys.length = t.length = t._values.length = t._items.length = 0;
+        t._keys.length = t.length = t._values.length = 0;
         return this;
     }
 
@@ -193,14 +220,13 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public delete(key:K):Boolean {
+    public delete(key: K): boolean {
         let t = this;
         let i = t._keys.indexOf(key);
         if (i > -1) {
             t._keys.splice(i, 1);
             t._values.splice(i, 1);
-            t._items.splice(i, 1);
-            t.length = t.size();
+            t.length = t._keys.length;
             return true;
         }
         return false;
@@ -213,12 +239,9 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public forEach(callbackfn:(value:V,key?:K,index?:number) => void):void {
-        let t = this;
-        let i = 0;
-        t._keys.forEach((v) => {
-            callbackfn(t.get(v),v,i);
-            i++;
+    public forEach(callbackfn: (value: V, key?: K, index?: number) => void): void {
+        this._keys.forEach((v, i) => {
+            callbackfn(this.get(v), v, i);
         });
     }
 
@@ -230,12 +253,9 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public map(callbackfn:(value:V,key?:K,index?:number) => any):Array<any> {
-        let t = this;
-        let i = -1;
-        return this.keys().map((itemKey) =>{
-            i++;
-            return callbackfn(t.get(itemKey),itemKey,i);
+    public map(callbackfn: (value: V, key?: K, index?: number) => any): any[] {
+        return this.keys().map((itemKey, i) => {
+            return callbackfn(this.get(itemKey), itemKey, i);
         });
     }
 
@@ -248,12 +268,10 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public filter(callbackfn:(value:V,key?:K,index?:number) => Boolean):TSMap<K,V> {
+    public filter(callbackfn: (value: V, key?: K, index?: number) => Boolean): this {
         let t = this;
-        let i = 0;
-        t._keys.forEach((v) => {
-            if(callbackfn(t.get(v),v,i) === false) t.delete(v); 
-            i++;
+        t._keys.forEach((v, i) => {
+            if (callbackfn(t.get(v), v, i) === false) t.delete(v);
         });
         return this;
     }
@@ -266,7 +284,7 @@ export class TSMap<K,V> {
      * 
      * @memberOf TSMap
      */
-    public clone():TSMap<K,V> {
-        return new TSMap<K,V>(<any> JSON.parse(JSON.stringify(this._items)));
+    public clone(): TSMap<K, V> {
+        return new TSMap<K, V>(this.entries());
     }
 }