+++ CHANGES.md
@@ -1,3 +1,11 @@
+# 0.0.4
+- Cleaned up code a bit.
+- Added `deepGet` property.
+- `toJSON` and `fromJSON` work with nested TSMaps now
+- Optimized several of the looping functions.
+- Upgraded to newest Typescript lib.
+- Updated readme.
+
 # 0.0.3
 - Cleaned up Readme.
 - First "official" release.